requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

# Core App
streamlit>=1.47.0
google-genai
pydantic

# Database
sqlalchemy
psycopg2-binary

# Document & Text Processing
numpy
xhtml2pdf

# RAG System
torch
faiss-cpu
sentence-transformers

# For local environment variables
python-dotenv